+++ Sources/SwiftATV/ATVClient.swift
@@ -8,7 +8,7 @@ import Foundation
 public enum ATVClient {
 
     /// Library version.
-    public static let version = "0.2.1"
+    public static let version = "0.2.2"
 
     fileprivate static let connectProtocolPriority: [ATVProtocol] = [
         .mrp, .companion, .airPlay, .raop, .dmap,